@@ -1,16 +1,12 @@
 {
   "name": "big.js",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "description": "",
   "main": "build/src/index.js",
   "typings": "build/src/index.d.ts",
   "keywords": [
-    "dfp",
-    "lib",
-    "api",
-    "ads",
-    "google",
-    "examples"
+    "big",
+    "int"
   ],
   "license": "MIT",
   "author": {
@@ -25,42 +21,52 @@
     "url": "https://github.com/havelessbemore/googleads-node-lib.git"
   },
   "scripts": {
-    "build": "npm run lint && npm run generate-package",
-    "distribute": "npm run build && cd ./build/src && npm publish",
+    "check": "gts check",
+    "clean": "gts clean",
+    "compile": "tsc -p .",
     "coverage": "tsc && istanbul cover _mocha build/test/**/*.js && remap-istanbul -i ./coverage/coverage.json -o ./coverage/lcov-report -t html",
-    "generate-package": "node build/scripts/makePackage.js",
-    "lint": "tsc && tslint 'src/**/*.ts'",
-    "prepublish": "npm run lint",
+    "fix": "gts fix",
+    "gen-1-compile": "npm run check && npm run compile",
+    "gen-2-webpack": "npm run webpack",
+    "gen-3-package": "node build/scripts/makePackage.js",
+    "gen-4-distribute": "cd ./build/src && npm publish",
+    "prepare": "npm run compile",
+    "pretest": "npm run compile",
+    "posttest": "npm run check",
     "start": "webpack-dev-server --hot --inline --progress --colors",
-    "test": "tsc && mocha --opts mocha.opts",
+    "test": "mocha --opts mocha.opts",
     "webpack": "webpack -p --display-error-details"
   },
   "dependencies": {
-    "babel-runtime": "^6.23.0"
+    "babel-runtime": "^6.26.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.0.0",
-    "@types/chalk": "^0.4.31",
-    "@types/mocha": "^2.2.41",
-    "@types/node": "^8.0.1",
+    "@types/chai": "^4.1.2",
+    "@types/chalk": "^2.2.0",
+    "@types/mocha": "^5.0.0",
+    "@types/node": "^9.6.2",
     "@types/rewire": "^2.5.28",
-    "@types/sinon": "^2.3.1",
-    "babel-core": "^6.25.0",
-    "babel-loader": "^7.1.0",
+    "@types/sinon": "^4.3.1",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.4",
     "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-babili": "^0.1.4",
-    "babel-preset-env": "^1.5.2",
-    "chai": "^4.0.2",
-    "chalk": "^1.1.3",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-minify": "^0.3.0",
+    "chai": "^4.1.2",
+    "chalk": "^2.3.2",
+    "gts": "^0.5.4",
     "istanbul": "^0.4.5",
-    "mocha": "^3.4.2",
-    "remap-istanbul": "^0.9.5",
-    "rewire": "^2.5.2",
-    "sinon": "^2.3.5",
-    "ts-loader": "^2.1.0",
-    "ts-node": "^3.0.6",
-    "tslint": "^5.4.3",
-    "typescript": "^2.3.4",
-    "webpack": "^3.0.0"
+    "mocha": "^5.0.5",
+    "remap-istanbul": "^0.11.0",
+    "rewire": "^3.0.2",
+    "sinon": "^4.5.0",
+    "ts-loader": "^4.1.0",
+    "ts-node": "^5.0.1",
+    "tslint": "^5.9.1",
+    "typescript": "^2.8.1",
+    "uglifyjs-webpack-plugin": "^1.2.4",
+    "webpack": "^4.5.0",
+    "webpack-bundle-analyzer": "^2.11.1",
+    "webpack-cli": "^2.0.14"
   }
 }
