@@ -9,7 +9,7 @@ export function singleDigitDivision(
   A: number[], minA: number, maxA: number,
   B: number, base: number
 ): [number, number] {
-  let remainder: number = 0;
+  let remainder = 0;
   for(let a: number = maxA; a-- > minA; remainder = remainder % B){
     remainder = remainder * base + A[a];
     A[a] = 0 | (remainder / B);
