@@ -26,14 +26,14 @@ import {toString} from '../functional/toString';
 import {assign} from '../util/intUtils';
 
 export class BigInt implements Integer {
-  public static get MINUS_ONE(): BigInt {return new BigInt(-1)};
-  public static get ZERO(): BigInt {return new BigInt(0)};
-  public static get ONE(): BigInt {return new BigInt(1)};
+  static get MINUS_ONE(): BigInt {return new BigInt(-1);}
+  static get ZERO(): BigInt {return new BigInt(0);}
+  static get ONE(): BigInt {return new BigInt(1);}
 
-  public base: number;
-  public digits: number[];
-  public precision: number;
-  public isNegative: boolean;
+  base!: number;
+  digits!: number[];
+  precision!: number;
+  isNegative!: boolean;
 
   constructor(input: Integer | number | string) {
     if(isInteger(input)){
@@ -43,144 +43,145 @@ export class BigInt implements Integer {
     }
   }
 
-  public abs(): BigInt {
+  abs(): BigInt {
     return BigInt.tryBigInt(abs(this));
-  };
+  }
 
-  public add(addend: Integer): BigInt {
+  add(addend: Integer): BigInt {
     return BigInt.tryBigInt(add(this, addend));
   }
 
-  public clone(): BigInt {
+  clone(): BigInt {
     return new BigInt(this);
   }
 
-  public compareTo(B: Integer): number {
+  compareTo(B: Integer): number {
     return compare(this, B);
   }
 
-  public set(source: Integer): BigInt {
-    return <BigInt>copy(this, source);
-  }
-
-  public divide(divisor: Integer): BigInt {
+  divide(divisor: Integer): BigInt {
     return this.divideAndRemainder(divisor)[0];
   }
 
-  public divideAndRemainder(divisor: Integer): [BigInt, BigInt] {
+  divideAndRemainder(divisor: Integer): [BigInt, BigInt] {
     const [Q, R]: [Integer, Integer] = divideAndRemainder(this, divisor);
-    return [BigInt.tryBigInt(Q), <BigInt>assign(BigInt.ZERO, R)];
+    return [BigInt.tryBigInt(Q), assign(BigInt.ZERO, R) as BigInt];
   }
 
-  public double(): BigInt {
+  double(): BigInt {
     return BigInt.tryBigInt(double(this));
   }
 
-  public gcd(B: Integer): BigInt {
+  gcd(B: Integer): BigInt {
     return BigInt.tryBigInt(gcd(this, B));
   }
 
-  public getBase(): number {
+  getBase(): number {
     return this.base;
   }
 
-  public halve(): [BigInt, BigInt] {
+  halve(): [BigInt, BigInt] {
     const [Q, R]: [Integer, Integer] = halve(this);
     return [BigInt.tryBigInt(Q), new BigInt(R.precision)];
   }
 
-  public static isBigInt(n: any): n is BigInt {
+  // tslint:disable-next-line:no-any
+  static isBigInt(n: any): n is BigInt {
     return n instanceof BigInt;
   }
 
-  public isEven(): boolean {
+  isEven(): boolean {
     return isEven(this);
   }
 
-  public isOdd(): boolean {
+  isOdd(): boolean {
     return !isEven(this);
   }
 
-  public lcm(B: Integer): BigInt {
+  lcm(B: Integer): BigInt {
     return BigInt.tryBigInt(lcm(this, B));
   }
 
-  public static max(A: Integer, B: Integer): BigInt {
+  static max(A: Integer, B: Integer): BigInt {
     return BigInt.tryBigInt(max(A, B));
   }
 
-  public max(B: Integer): BigInt {
+  max(B: Integer): BigInt {
     const C: Integer = max(this, B);
     if(C === this){
       return this;
     }
     if(C === B){
-      return <BigInt>copy(this, C);
+      return copy(this, C) as BigInt;
     }
-    return <BigInt>assign(BigInt.ZERO, C);
+    return assign(BigInt.ZERO, C) as BigInt;
   }
 
-  public static min(A: Integer, B: Integer): BigInt {
+  static min(A: Integer, B: Integer): BigInt {
     return BigInt.tryBigInt(min(A, B));
   }
 
-  public min(B: Integer): BigInt {
+  min(B: Integer): BigInt {
     const C: Integer = min(this, B);
     if(C === this){
       return this;
     }
     if(C === B){
-      return <BigInt>copy(this, C);
+      return copy(this, C) as BigInt;
     }
-    return <BigInt>assign(BigInt.ZERO, C);
+    return assign(BigInt.ZERO, C) as BigInt;
   }
 
-  public minusminus(): BigInt {
+  minusminus(): BigInt {
     return BigInt.tryBigInt(decrement(this));
   }
 
-  public multiply(multiplier: Integer): BigInt {
+  multiply(multiplier: Integer): BigInt {
     return BigInt.tryBigInt(multiply(this, multiplier));
   }
 
-  public negate(): BigInt {
+  negate(): BigInt {
     return BigInt.tryBigInt(negate(this));
-  };
+  }
 
-  public plusplus(): BigInt {
+  plusplus(): BigInt {
     return BigInt.tryBigInt(increment(this));
   }
 
-  public pow(power: Integer): BigInt {
+  pow(power: Integer): BigInt {
     return BigInt.tryBigInt(pow(this, power));
   }
 
-  public remainder(divisor: Integer): BigInt {
+  remainder(divisor: Integer): BigInt {
     const R: BigInt = this.divideAndRemainder(divisor)[1];
-    return <BigInt>assign(this, R);
+    return assign(this, R) as BigInt;
   }
 
-  public setBase(base: number): BigInt {
+  set(source: Integer): BigInt {
+    return copy(this, source) as BigInt;
+  }
+
+  setBase(base: number): BigInt {
     return BigInt.tryBigInt(setBase(this, base));
   }
 
-  public signum(): number {
+  signum(): number {
     return signum(this);
-  };
+  }
 
-  public square(): BigInt {
+  square(): BigInt {
     return BigInt.tryBigInt(square(this));
   }
 
-  public subtract(subtrahend: Integer): BigInt {
+  subtract(subtrahend: Integer): BigInt {
     return BigInt.tryBigInt(subtract(this, subtrahend));
   }
 
-  public toString(): string {
+  toString(): string {
     return toString(this);
   }
 
   private static tryBigInt(A: Integer): BigInt {
-    return BigInt.isBigInt(A) ? A : <BigInt>assign(BigInt.ZERO, A);
+    return BigInt.isBigInt(A) ? A : assign(BigInt.ZERO, A) as BigInt;
   }
 }
