@@ -7,6 +7,7 @@ export function copy<T>(to: T[], minTo: number, from: ReadonlyArray<T>, minFrom:
   }
 }
 
+// tslint:disable-next-line:no-any
 export function growArray(A: any[], minNewLen: number, maxNewLen: number): void {
   let newLen: number = maxNewLen;
   if(newLen > Globals.MAX_PRECISION){
@@ -23,8 +24,9 @@ export function growArray(A: any[], minNewLen: number, maxNewLen: number): void 
   }
 }
 
+// tslint:disable-next-line:no-any
 export function printArr(A: any[], min: number, max: number, minSub: number = min, maxSub: number = max): string {
-  A = A.map(function(v){return (v == null || v !== v) ? v + '' : v;});
+  A = A.map(v => (v == null || v !== v) ? v + '' : v);
   return "[" +
     A.slice(min,minSub).join(', ') +
     "|" + A.slice(minSub,maxSub).join(', ') +
@@ -32,12 +34,14 @@ export function printArr(A: any[], min: number, max: number, minSub: number = mi
   "]";
 }
 
+// tslint:disable-next-line:no-any
 export function safeShiftUp(A: any[], min: number, max: number, shifts: number): void {
   for(let i: number = max + shifts; max > min; A[--i] = A[--max]){
   }
 }
 
 //Assumes shifts >= max - min, shifts > 0
+// tslint:disable-next-line:no-any
 export function unsafeShiftUp(A: any[], min: number, max: number, shifts: number): void {
   for(let i: number = min + shifts; min < max; A[i++] = A[min++]){
   }
