@@ -7,45 +7,45 @@ function test(A: number[], min: number, maxA: number, B: number[], maxB: number,
   assert.deepEqual(A, B);
 }
 
-describe('decrement', function(){
+describe('decrement', () => {
 
-  it('should return 1-1 in base 10', function(){
+  it('should return 1-1 in base 10', () => {
     test([1], 0, 1, [0], 0, 10);
   });
 
-  it('should return 2-1 in base 10', function(){
+  it('should return 2-1 in base 10', () => {
     test([2], 0, 1, [1], 1, 10);
   });
 
-  it('should return 10-1 in base 10', function(){
+  it('should return 10-1 in base 10', () => {
     test([0,1], 0, 2, [9, 0], 1, 10);
   });
 
-  it('should return 11-1 in base 10', function(){
+  it('should return 11-1 in base 10', () => {
     test([1,1], 0, 2, [0,1], 2, 10);
   });
 
-  it('should return 99-1 in base 10', function(){
+  it('should return 99-1 in base 10', () => {
     test([9,9], 0, 2, [8,9], 2, 10);
   });
 
-  it('should return 100000-1 in base 10', function(){
+  it('should return 100000-1 in base 10', () => {
     test([0,0,0,0,0,1], 0, 6, [9,9,9,9,9,0], 5, 10);
   });
 
-  it('should return 300000-1 in base 10', function(){
+  it('should return 300000-1 in base 10', () => {
     test([0,0,0,0,0,3], 0, 6, [9,9,9,9,9,2], 6, 10);
   });
 
-  it('should return 321-1 in base 179', function(){
+  it('should return 321-1 in base 179', () => {
     test([142, 1], 0, 2, [141, 1], 2, 179);
   });
 
-  it('should return 28,676,695-1 in base 179', function(){
+  it('should return 28,676,695-1 in base 179', () => {
     test([0,0,0,5], 0, 4, [178,178,178,4], 4, 179);
   });
 
-  it('should return 72-1 in base 2', function(){
+  it('should return 72-1 in base 2', () => {
     test([0,0,0,1,0,0,1], 0, 7, [1,1,1,0,0,0,1], 7, 2);
   });
 });
