@@ -10,29 +10,29 @@ function test(A: number[], minA: number, maxA: number, B: number[], minB: number
   assert.deepEqual(A, C);
 }
 
-describe('longMultiplication', function(){
+describe('longMultiplication', () => {
 
-  it('should return 1*1 in base 10', function(){
+  it('should return 1*1 in base 10', () => {
     test([1], 0, 1, [1], 0, 1, [1], 1, 10);
   });
 
-  it('should return 1*9 in base 10', function(){
+  it('should return 1*9 in base 10', () => {
     test([1], 0, 1, [9], 0, 1, [9], 1, 10);
   });
 
-  it('should return 9*8 in base 10', function(){
+  it('should return 9*8 in base 10', () => {
     test([9], 0, 1, [8], 0, 1, [2,7], 2, 10);
   });
 
-  it('should return 8*9 in base 10', function(){
+  it('should return 8*9 in base 10', () => {
     test([8], 0, 1, [9], 0, 1, [2,7], 2, 10);
   });
 
-  it('should return 123 * 456789 in base 10', function(){
+  it('should return 123 * 456789 in base 10', () => {
     test([3,2,1], 0, 3, [9,8,7,6,5,4], 0, 6, [7,4,0,5,8,1,6,5], 8, 10);
   });
 
-  it('should return 456789 * 123 in base 2', function(){
+  it('should return 456789 * 123 in base 2', () => {
     test(
       [1,0,1,0,1,0,1,0,0,0,0,1,1,1,1,1,0,1,1], 0, 19,
       [1,1,0,1,1,1,1], 0, 7,
@@ -40,7 +40,7 @@ describe('longMultiplication', function(){
     );
   });
 
-  it('should return 987654 * 321 in base 179', function(){
+  it('should return 987654 * 321 in base 179', () => {
     test(
       [111, 147, 30], 0, 3,
       [142, 1], 0, 2,
@@ -48,7 +48,7 @@ describe('longMultiplication', function(){
     );
   });
 
-  it('should return 1234567 * 1234567 in base 10', function(){
+  it('should return 1234567 * 1234567 in base 10', () => {
     test(
       [7,6,5,4,3,2,1], 0, 7,
       [7,6,5,4,3,2,1], 0, 7,
@@ -56,7 +56,7 @@ describe('longMultiplication', function(){
     );
   });
 
-  it('should return 1234567 * 1,524,155,677,489 in base 10', function(){
+  it('should return 1234567 * 1,524,155,677,489 in base 10', () => {
     test(
       [7,6,5,4,3,2,1], 0, 7,
       [9,8,4,7,7,6,5,5,1,4,2,5,1], 0, 13,
@@ -64,7 +64,7 @@ describe('longMultiplication', function(){
     );
   });
 
-  it('should return 9,870,006,540,003,210 * 123,000,456,000,789 in base 10', function(){
+  it('should return 9,870,006,540,003,210 * 123,000,456,000,789 in base 10', () => {
     test(
       [0,1,2,3,0,0,0,4,5,6,0,0,0,7,8,9], 0, 16,
       [9,8,7,0,0,0,6,5,4,0,0,0,3,2,1], 0, 15,
@@ -72,7 +72,7 @@ describe('longMultiplication', function(){
     );
   });
 
-  it('should return 123,000,456,000,789 * 9,870,006,540,003,210 in base 10', function(){
+  it('should return 123,000,456,000,789 * 9,870,006,540,003,210 in base 10', () => {
     test(
       [9,8,7,0,0,0,6,5,4,0,0,0,3,2,1], 0, 15,
       [0,1,2,3,0,0,0,4,5,6,0,0,0,7,8,9], 0, 16,
@@ -80,7 +80,7 @@ describe('longMultiplication', function(){
     );
   });
 
-  it('should return 1,524,155,677,489 * 1,524,155,677,489 in base 10', function(){
+  it('should return 1,524,155,677,489 * 1,524,155,677,489 in base 10', () => {
       test(
         [9,8,4,7,7,6,5,5,1,4,2,5,1], 0, 13,
         [9,8,4,7,7,6,5,5,1,4,2,5,1], 0, 13,
