@@ -10,29 +10,29 @@ function test(A: number[], B: number[], C: number[], base: number): void {
   assert.deepEqual(A, C);
 }
 
-describe('subtraction', function(){
+describe('subtraction', () => {
 
-  it('should return 0-0 in base 10', function(){
+  it('should return 0-0 in base 10', () => {
     test([0], [0], [], 10);
   });
 
-  it('should return 1-1 in base 10', function(){
+  it('should return 1-1 in base 10', () => {
     test([1], [1], [], 10);
   });
 
-  it('should return 10-10 in base 10', function(){
+  it('should return 10-10 in base 10', () => {
     test([0,1], [0,1], [], 10);
   });
 
-  it('should return 9-1 in base 10', function(){
+  it('should return 9-1 in base 10', () => {
     test([9], [1], [8], 10);
   });
 
-  it('should return 456789-123 in base 10', function(){
+  it('should return 456789-123 in base 10', () => {
     test([9,8,7,6,5,4], [3,2,1], [6,6,6,6,5,4], 10);
   });
 
-  it('should return 456789-123 in base 2', function(){
+  it('should return 456789-123 in base 2', () => {
     test(
       [1,0,1,0,1,0,1,0,0,0,0,1,1,1,1,1,0,1,1],
       [1,1,0,1,1,1,1],
@@ -40,7 +40,7 @@ describe('subtraction', function(){
     );
   });
 
-  it('should return 9,870,006,540,003,210-123,000,456,000,789 in base 10', function(){
+  it('should return 9,870,006,540,003,210-123,000,456,000,789 in base 10', () => {
     test(
       [0,1,2,3,0,0,0,4,5,6,0,0,0,7,8,9],
       [9,8,7,0,0,0,6,5,4,0,0,0,3,2,1],
@@ -48,7 +48,7 @@ describe('subtraction', function(){
     );
   });
 
-  it('should return 1,524,155,677,489-1,524,155,677,489 in base 10', function(){
+  it('should return 1,524,155,677,489-1,524,155,677,489 in base 10', () => {
     test(
       [9,8,4,7,7,6,5,5,1,4,2,5,1],
       [9,8,4,7,7,6,5,5,1,4,2,5,1],
@@ -56,7 +56,7 @@ describe('subtraction', function(){
     );
   });
 
-  it('should return 10,000,000-1 in base 10', function(){
+  it('should return 10,000,000-1 in base 10', () => {
     test(
       [0,0,0,0,0,0,0,1],
       [1],
@@ -64,7 +64,7 @@ describe('subtraction', function(){
     );
   });
 
-  it('should return 10,100,000-1 in base 10', function(){
+  it('should return 10,100,000-1 in base 10', () => {
     test(
       [0,0,0,0,0,1,0,1],
       [1],
