@@ -3,8 +3,8 @@ import {Integer} from '../../src/integer';
 import {abs} from '../../src/functional/abs';
 import {toInteger} from '../../src/util/intUtils';
 
-describe('abs', function(){
-  it('should return the absolute value of a positive input', function(){
+describe('abs', () => {
+  it('should return the absolute value of a positive input', () => {
     const input: Integer = toInteger([1,2,3,4,5], 5, false, 125);
     const output: Integer = abs(input);
     assert.equal(input, output);
@@ -12,7 +12,7 @@ describe('abs', function(){
     assert.equal(output.isNegative, false);
   });
 
-  it('should return the absolute value of a negative input', function(){
+  it('should return the absolute value of a negative input', () => {
     const input: Integer = toInteger([1,2,3,4,5], 5, true, 125);
     const output: Integer = abs(input);
     assert.equal(input, output);
