@@ -3,9 +3,9 @@ import {Integer} from '../../src/integer';
 import {lcm} from '../../src/functional/lcm';
 import {toInteger} from '../../src/util/intUtils';
 
-describe('lcm', function(){
+describe('lcm', () => {
 
-  it('should return lcm(A,A) = A in base 10', function(){
+  it('should return lcm(A,A) = A in base 10', () => {
     const A: Integer = toInteger([3,2,1], 3, false, 10);
     const expected: Integer = toInteger([3,2,1], 3, false, 10);
 
@@ -18,7 +18,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(0,B) = 0 in base 10', function(){
+  it('should return lcm(0,B) = 0 in base 10', () => {
     const A: Integer = toInteger([], 0, false, 10);
     const B: Integer = toInteger([3,2,1], 3, false, 10);
     const expected: Integer = toInteger([], 0, false, 10);
@@ -32,7 +32,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(A,1) = A in base 10', function(){
+  it('should return lcm(A,1) = A in base 10', () => {
     const A: Integer = toInteger([3,2,1], 3, false, 10);
     const B: Integer = toInteger([1], 1, false, 10);
     const expected: Integer = toInteger([3,2,1], 3, false, 10);
@@ -46,7 +46,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(A,0) = 0 in base 10', function(){
+  it('should return lcm(A,0) = 0 in base 10', () => {
     const A: Integer = toInteger([3,2,1], 3, false, 10);
     const B: Integer = toInteger([], 0, false, 10);
     const expected: Integer = toInteger([], 0, false, 10);
@@ -60,7 +60,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(1,B) = B in base 10', function(){
+  it('should return lcm(1,B) = B in base 10', () => {
     const A: Integer = toInteger([1], 1, false, 10);
     const B: Integer = toInteger([3,2,1], 3, false, 10);
     const expected: Integer = toInteger([3,2,1], 3, false, 10);
@@ -74,7 +74,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(1,1) = 1 in base 10', function(){
+  it('should return lcm(1,1) = 1 in base 10', () => {
     const A: Integer = toInteger([1], 1, false, 10);
     const B: Integer = toInteger([1], 1, false, 10);
     const expected: Integer = toInteger([1], 1, false, 10);
@@ -88,7 +88,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(4,6) = 12 in base 10', function(){
+  it('should return lcm(4,6) = 12 in base 10', () => {
     const A: Integer = toInteger([4], 1, false, 10);
     const B: Integer = toInteger([6], 1, false, 10);
     const expected: Integer = toInteger([2,1], 2, false, 10);
@@ -102,7 +102,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(21,6) = 42 in base 10', function(){
+  it('should return lcm(21,6) = 42 in base 10', () => {
     const A: Integer = toInteger([1,2], 2, false, 10);
     const B: Integer = toInteger([6], 1, false, 10);
     const expected: Integer = toInteger([2,4], 2, false, 10);
@@ -116,7 +116,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(21,6) = 42 in base 2', function(){
+  it('should return lcm(21,6) = 42 in base 2', () => {
     const A: Integer = toInteger([1,0,1,0,1], 5, false, 2);
     const B: Integer = toInteger([0,1,1], 3, false, 2);
     const expected: Integer = toInteger([0,1,0,1,0,1], 6, false, 2);
@@ -130,7 +130,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(21,6) = 42 in base 3', function(){
+  it('should return lcm(21,6) = 42 in base 3', () => {
     const A: Integer = toInteger([0,1,2], 3, false, 3);
     const B: Integer = toInteger([0,2], 2, false, 3);
     const expected: Integer = toInteger([0,2,1,1], 4, false, 3);
@@ -144,7 +144,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(21,6) = 42 in base 5', function(){
+  it('should return lcm(21,6) = 42 in base 5', () => {
     const A: Integer = toInteger([1,4], 2, false, 5);
     const B: Integer = toInteger([1,1], 2, false, 5);
     const expected: Integer = toInteger([2,3,1], 3, false, 5);
@@ -158,7 +158,7 @@ describe('lcm', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return lcm(21,6) = 42 in base 16', function(){
+  it('should return lcm(21,6) = 42 in base 16', () => {
     const A: Integer = toInteger([5,1], 2, false, 16);
     const B: Integer = toInteger([6], 1, false, 16);
     const expected: Integer = toInteger([10,2], 2, false, 16);
