@@ -10,22 +10,22 @@ interface MaxFunc {
   max: (A: Integer, B: Integer, b?: boolean) => Integer;
 }
 
-describe('max', function(){
-  it('should return A as max', function(){
+describe('max', () => {
+  it('should return A as max', () => {
     const A: Integer = toInteger([1,2,3,4,5], 5, false, 125);
     const B: Integer = toInteger([6,7,8], 3, false, 10);
     const C: Integer = max(A, B);
     assert.equal(A, C);
   });
 
-  it('should return B as max', function(){
+  it('should return B as max', () => {
     const A: Integer = toInteger([1,2], 2, false, 10);
     const B: Integer = toInteger([3,4], 2, false, 10);
     const C: Integer = max(A, B);
     assert.equal(B, C);
   });
 
-  it('should use compare() to find max', function(){
+  it('should use compare() to find max', () => {
     const A: Integer = toInteger([1,2], 2, false, 10);
     const B: Integer = toInteger([3,4], 2, false, 10);
 
