@@ -3,9 +3,9 @@ import {Integer} from '../../src/integer';
 import {multiply} from '../../src/functional/multiply';
 import {toInteger} from '../../src/util/intUtils';
 
-describe('multiply', function(){
+describe('multiply', () => {
 
-  it('should return A*A in base 10', function(){
+  it('should return A*A in base 10', () => {
     const A: Integer = toInteger([2], 1, false, 10);
     const expected: Integer = toInteger([4], 1, false, 10);
 
@@ -18,7 +18,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 0*123 in base 10', function(){
+  it('should return 0*123 in base 10', () => {
     const A: Integer = toInteger([], 0, false, 10);
     const B: Integer = toInteger([3,2,1], 3, false, 10);
     const expected: Integer = toInteger([], 0, false, 10);
@@ -32,7 +32,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 123*0 in base 10', function(){
+  it('should return 123*0 in base 10', () => {
     const A: Integer = toInteger([3,2,1], 3, false, 10);
     const B: Integer = toInteger([], 0, false, 10);
     const expected: Integer = toInteger([], 0, false, 10);
@@ -46,7 +46,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 1*1 in base 10', function(){
+  it('should return 1*1 in base 10', () => {
       const A: Integer = toInteger([1], 1, false, 10);
       const B: Integer = toInteger([1], 1, false, 10);
       const expected: Integer = toInteger([1], 1, false, 10);
@@ -60,7 +60,7 @@ describe('multiply', function(){
       assert.deepEqual(actual, expected);
   });
 
-  it('should return 1*10 in base 10', function(){
+  it('should return 1*10 in base 10', () => {
     const A: Integer = toInteger([1], 1, false, 10);
     const B: Integer = toInteger([0,1], 2, false, 10);
     const expected: Integer = toInteger([0,1], 2, false, 10);
@@ -74,7 +74,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 1*100 in base 10', function(){
+  it('should return 1*100 in base 10', () => {
     const A: Integer = toInteger([1], 1, false, 10);
     const B: Integer = toInteger([0,0,1], 3, false, 10);
     const expected: Integer = toInteger([0,0,1], 3, false, 10);
@@ -88,7 +88,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 9*8 in base 10', function(){
+  it('should return 9*8 in base 10', () => {
     const A: Integer = toInteger([9], 1, false, 10);
     const B: Integer = toInteger([8], 1, false, 10);
     const expected: Integer = toInteger([2,7], 2, false, 10);
@@ -102,7 +102,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 8*9 in base 10', function(){
+  it('should return 8*9 in base 10', () => {
     const A: Integer = toInteger([8], 1, false, 10);
     const B: Integer = toInteger([9], 1, false, 10);
     const expected: Integer = toInteger([2,7], 2, false, 10);
@@ -116,7 +116,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 123 * 2 in base 10', function(){
+  it('should return 123 * 2 in base 10', () => {
     const A: Integer = toInteger([3,2,1], 3, false, 10);
     const B: Integer = toInteger([2], 1, false, 10);
     const expected: Integer = toInteger([6,4,2], 3, false, 10);
@@ -130,7 +130,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 2 * 123 in base 10', function(){
+  it('should return 2 * 123 in base 10', () => {
     const A: Integer = toInteger([2], 1, false, 10);
     const B: Integer = toInteger([3,2,1], 3, false, 10);
     const expected: Integer = toInteger([6,4,2], 3, false, 10);
@@ -144,7 +144,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 3 * 123 in base 10', function(){
+  it('should return 3 * 123 in base 10', () => {
     const A: Integer = toInteger([3], 1, false, 10);
     const B: Integer = toInteger([3,2,1], 3, false, 10);
     const expected: Integer = toInteger([9,6,3], 3, false, 10);
@@ -158,7 +158,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 123 * 456789 in base 10', function(){
+  it('should return 123 * 456789 in base 10', () => {
     const A: Integer = toInteger([3,2,1], 3, false, 10);
     const B: Integer = toInteger([9,8,7,6,5,4], 6, false, 10);
     const expected: Integer = toInteger([7,4,0,5,8,1,6,5], 8, false, 10);
@@ -172,7 +172,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return 456789 * 123 in base 2', function(){
+  it('should return 456789 * 123 in base 2', () => {
     const A: Integer = toInteger([1,0,1,0,1,0,1,0,0,0,0,1,1,1,1,1,0,1,1], 19, false, 2);
     const B: Integer = toInteger([1,1,0,1,1,1,1], 7, false, 2);
     const expected: Integer = toInteger([1,1,1,0,1,0,1,1,0,0,0,0,1,0,1,0,1,0,0,1,1,0,1,0,1,1], 26, false, 2);
@@ -186,7 +186,7 @@ describe('multiply', function(){
     assert.deepEqual(actual, expected);
   });
 
-  it('should return [100 digit number] ^ [100 digit number] in base 10', function(){
+  it('should return [100 digit number] ^ [100 digit number] in base 10', () => {
     const A: Integer = toInteger([
       6,9,5,9,8,3,4,7,9,9,2,4,9,1,1,6,6,1,8,2,9,0,8,2,4,2,6,2,4,2,
       5,3,5,9,3,1,0,1,1,6,2,7,6,7,1,9,8,7,5,8,4,2,7,9,7,2,8,2,3,4,
