@@ -3,8 +3,8 @@ import {Integer} from '../../src/integer';
 import {toInteger} from '../../src/util/intUtils';
 import {negate} from '../../src/functional/negate';
 
-describe('negate', function(){
-  it('should return the negative value of a positive input', function(){
+describe('negate', () => {
+  it('should return the negative value of a positive input', () => {
     const input: Integer = toInteger([1,2,3,4,5], 5, false, 125);
     const output: Integer = negate(input);
     assert.equal(input, output);
@@ -12,7 +12,7 @@ describe('negate', function(){
     assert.deepEqual(input, output);
   });
 
-  it('should return the negative value of a negative input', function(){
+  it('should return the negative value of a negative input', () => {
     const input: Integer = toInteger([1,2,3,4,5], 5, true, 125);
     const output: Integer = negate(input);
     assert.equal(input, output);
