@@ -4,10 +4,10 @@ import {
   copy, growArray, printArr, safeShiftUp, unsafeShiftUp
 } from '../../src/util/arrayUtils';
 
-describe('arrayUtils', function(){
+describe('arrayUtils', () => {
 
-  describe('copy', function(){
-    it('should copy indices [min, max) of source array into target array', function(){
+  describe('copy', () => {
+    it('should copy indices [min, max) of source array into target array', () => {
       const source: number[] = [1,2,3,4,5,6,7,8,9];
       const target: number[] = [0,0,0,0,0,0,0,0,0];
       copy(target, 1, source, 4, 4);
@@ -23,8 +23,8 @@ describe('arrayUtils', function(){
     });
   });
 
-  describe('growArray', function(){
-    it('should not change size if new size <= current size', function(){
+  describe('growArray', () => {
+    it('should not change size if new size <= current size', () => {
       const A: number[] = [0,1,2,3,4];
       growArray(A, 2, 4);
       assert.equal(A.length, 5);
@@ -34,27 +34,27 @@ describe('arrayUtils', function(){
       assert.equal(A.length, 5);
     });
 
-    it('should change to max size if valid and > current size', function(){
+    it('should change to max size if valid and > current size', () => {
       const A: number[] = [0,1,2,3,4];
       growArray(A, 2, 10);
       assert.equal(A.length, 10);
     });
 
-    it('should change to min size if > current size and max size invalid', function(){
+    it('should change to min size if > current size and max size invalid', () => {
       const A: number[] = [0,1,2,3,4];
       growArray(A, 10, Globals.MAX_PRECISION + 1);
       assert.equal(A.length, 10);
     });
 
-    it('should throw error if new sizes > supported sizes', function(){
+    it('should throw error if new sizes > supported sizes', () => {
       const A: number[] = [0,1,2,3,4];
       const max: number = Globals.MAX_PRECISION;
-      expect(function(){growArray(A, max + 1, max + 1)}).to.throw(RangeError);
+      expect(() => growArray(A, max + 1, max + 1)).to.throw(RangeError);
     });
   });
 
-  describe('printArr', function(){
-    it("should print out the array in subsets", function(){
+  describe('printArr', () => {
+    it("should print out the array in subsets", () => {
       assert.equal(printArr([],0,0), "[||]");
       assert.equal(printArr([1,2,3],0,3), "[|1, 2, 3|]");
       assert.equal(printArr([1,2,3],0,2), "[|1, 2|]");
@@ -63,13 +63,13 @@ describe('arrayUtils', function(){
       assert.equal(printArr([1,2,3,4,5],0,5,2,3), "[1, 2|3|4, 5]");
     });
 
-    it("should print null / NaN / undefined correctly", function(){
+    it("should print null / NaN / undefined correctly", () => {
       assert.equal(printArr([null, undefined, NaN],0,3), "[|null, undefined, NaN|]");
     });
   });
 
-  describe('safeShiftUp', function(){
-    it("should correctly shift array segment up when no overlap with target segment", function(){
+  describe('safeShiftUp', () => {
+    it("should correctly shift array segment up when no overlap with target segment", () => {
       const actual: number[] = [0,1,2,3,4,7,7,7,7,7];
       safeShiftUp(actual, 3, 5, 4);
       assert.deepEqual(actual, [0,1,2,3,4,7,7,3,4,7]);
@@ -83,7 +83,7 @@ describe('arrayUtils', function(){
       assert.deepEqual(actual, [0,1,2,3,4,7,1,2,3,4,7]);
     });
 
-    it("should correctly shift array segment up when overlap with target segment", function(){
+    it("should correctly shift array segment up when overlap with target segment", () => {
       const actual: number[] = [0,1,2,3,4,7,1,2,3,4,7];
       safeShiftUp(actual, 1, 5, 1);
       assert.deepEqual(actual, [0,1,1,2,3,4,1,2,3,4,7]);
@@ -96,8 +96,8 @@ describe('arrayUtils', function(){
     });
   });
 
-  describe('unsafeShiftUp', function(){
-    it("should correctly shift array segment up when no overlap with target segment", function(){
+  describe('unsafeShiftUp', () => {
+    it("should correctly shift array segment up when no overlap with target segment", () => {
       const actual: number[] = [0,1,2,3,4,7,7,7,7,7];
       unsafeShiftUp(actual, 3, 5, 4);
       assert.deepEqual(actual, [0,1,2,3,4,7,7,3,4,7]);
@@ -111,7 +111,7 @@ describe('arrayUtils', function(){
       assert.deepEqual(actual, [0,1,2,3,4,7,1,2,3,4,7]);
     });
 
-    it("should incorrectly shift array segment up when overlap with target segment", function(){
+    it("should incorrectly shift array segment up when overlap with target segment", () => {
       const actual: number[] = [0,1,2,3,4,7,1,2,3,4,7];
       unsafeShiftUp(actual, 1, 5, 1);
       assert.deepEqual(actual, [0,1,1,1,1,1,1,2,3,4,7]);
